renamed folder and package from "aws"
